backend/requirements.txt
celery==5.3.6
fastapi==0.110.0
uvicorn
httpx==0.27.0
pydantic==1.10.2
pytz==2022.7
Requests==2.31.0
python-multipart
sqlalchemy
pymysql
sqlalchemy_utils
psycopg2
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.1.7
mysql-connector-python
tenacity
cryptography
typing_extensions
pydantic[email]
aioredis
hiredis
boto3
aioboto3
pytest
pytest-asyncio